refactoring, formatting
